+++ tests/CircleTest.php
@@ -1,0 +1,469 @@
+<?php
+
+declare(strict_types=1);
+
+namespace UIAwesome\Html\Svg\Tests;
+
+use PHPUnit\Framework\Attributes\Group;
+use PHPUnit\Framework\TestCase;
+use UIAwesome\Html\Core\Factory\SimpleFactory;
+use UIAwesome\Html\Core\Values\Aria;
+use UIAwesome\Html\Core\Values\DataProperty;
+use UIAwesome\Html\Core\Values\Language;
+use UIAwesome\Html\Core\Values\Role;
+use UIAwesome\Html\Svg\Circle;
+use UIAwesome\Html\Svg\Tests\Support\Stub\DefaultProvider;
+use UIAwesome\Html\Svg\Tests\Support\TestSupport;
+use UIAwesome\Html\Svg\Values\FillRule;
+use UIAwesome\Html\Svg\Values\StrokeLineCap;
+use UIAwesome\Html\Svg\Values\StrokeLineJoin;
+
+#[Group('svg')]
+final class CircleTest extends TestCase
+{
+    use TestSupport;
+
+    public function testRenderWithAddAriaAttribute(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <circle aria-pressed="true">
+            HTML,
+            Circle::tag()->addAriaAttribute('pressed', true)->render(),
+            "Failed asserting that element renders correctly with 'addAriaAttribute()' method.",
+        );
+    }
+
+    public function testRenderWithAddAriaAttributeUsingEnum(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <circle aria-pressed="true">
+            HTML,
+            Circle::tag()->addAriaAttribute(Aria::PRESSED, true)->render(),
+            "Failed asserting that element renders correctly with 'addAriaAttribute()' method.",
+        );
+    }
+
+    public function testRenderWithAddDataAttribute(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <circle data-value="value">
+            HTML,
+            Circle::tag()->addDataAttribute('value', 'value')->render(),
+            "Failed asserting that element renders correctly with 'addDataAttribute()' method.",
+        );
+    }
+
+    public function testRenderWithAddDataAttributeUsingEnum(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <circle data-value="value">
+            HTML,
+            Circle::tag()->addDataAttribute(DataProperty::VALUE, 'value')->render(),
+            "Failed asserting that element renders correctly with 'addDataAttribute()' method.",
+        );
+    }
+
+    public function testRenderWithAriaAttributes(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <circle aria-controls="modal-1" aria-hidden="false" aria-label="Close">
+            HTML,
+            Circle::tag()
+                ->ariaAttributes(
+                    [
+                        'controls' => static fn(): string => 'modal-1',
+                        'hidden' => false,
+                        'label' => 'Close',
+                    ],
+                )
+                ->render(),
+            "Failed asserting that element renders correctly with 'ariaAttributes()' method.",
+        );
+    }
+
+    public function testRenderWithAttributes(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <circle class="value">
+            HTML,
+            Circle::tag()->attributes(['class' => 'value'])->render(),
+            "Failed asserting that element renders correctly with 'attributes()' method.",
+        );
+    }
+
+    public function testRenderWithClass(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <circle class="value">
+            HTML,
+            Circle::tag()->class('value')->render(),
+        );
+    }
+
+    public function testRenderWithDataAttributes(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <circle data-value="test-value">
+            HTML,
+            Circle::tag()->dataAttributes(['value' => 'test-value'])->render(),
+            "Failed asserting that element renders correctly with 'dataAttributes()' method.",
+        );
+    }
+
+    public function testRenderWithDefaultConfigurationValues(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <circle class="default-class">
+            HTML,
+            Circle::tag(['class' => 'default-class'])->render(),
+            'Failed asserting that default configuration values are applied correctly.',
+        );
+    }
+
+    public function testRenderWithDefaultProvider(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <circle class="default-class" title="default-title">
+            HTML,
+            Circle::tag()->addDefaultProvider(DefaultProvider::class)->render(),
+            'Failed asserting that default provider is applied correctly.',
+        );
+    }
+
+    public function testRenderWithFill(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <circle fill="#ff0000">
+            HTML,
+            Circle::tag()->fill('#ff0000')->render(),
+            "Failed asserting that element renders correctly with 'fill' attribute.",
+        );
+    }
+
+    public function testRenderWithFillOpacity(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <circle fill-opacity="0.7">
+            HTML,
+            Circle::tag()->fillOpacity('0.7')->render(),
+            "Failed asserting that element renders correctly with 'fill-opacity' attribute.",
+        );
+    }
+
+    public function testRenderWithFillRule(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <circle fill-rule="evenodd">
+            HTML,
+            Circle::tag()->fillRule('evenodd')->render(),
+            "Failed asserting that element renders correctly with 'fill-rule' attribute.",
+        );
+    }
+
+    public function testRenderWithFillRuleUsingEnum(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <circle fill-rule="nonzero">
+            HTML,
+            Circle::tag()->fillRule(FillRule::NONZERO)->render(),
+            "Failed asserting that element renders correctly with 'fill-rule' attribute.",
+        );
+    }
+
+    public function testRenderWithGlobalDefaultsAreApplied(): void
+    {
+        SimpleFactory::setDefaults(Circle::class, ['class' => 'default-class']);
+
+        self::equalsWithoutLE(
+            <<<HTML
+            <circle class="default-class">
+            HTML,
+            Circle::tag()->render(),
+            'Failed asserting that global defaults are applied correctly.',
+        );
+
+        SimpleFactory::setDefaults(Circle::class, []);
+    }
+
+    public function testRenderWithId(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <circle id="test-id">
+            HTML,
+            Circle::tag()->id('test-id')->render(),
+            "Failed asserting that element renders correctly with 'id' attribute.",
+        );
+    }
+
+    public function testRenderWithLangUsingEnum(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <circle lang="es">
+            HTML,
+            Circle::tag()->lang(Language::SPANISH)->render(),
+            "Failed asserting that element renders correctly with 'lang' attribute.",
+        );
+    }
+
+    public function testRenderWithOpacity(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <circle opacity="0.5">
+            HTML,
+            Circle::tag()->opacity('0.5')->render(),
+            "Failed asserting that element renders correctly with 'opacity' attribute.",
+        );
+    }
+
+    public function testRenderWithRole(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <circle role="banner">
+            HTML,
+            Circle::tag()->role('banner')->render(),
+            "Failed asserting that element renders correctly with 'role' attribute.",
+        );
+    }
+
+    public function testRenderWithRoleUsingEnum(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <circle role="banner">
+            HTML,
+            Circle::tag()->role(Role::BANNER)->render(),
+            "Failed asserting that element renders correctly with 'role' attribute.",
+        );
+    }
+
+    public function testRenderWithStroke(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <circle stroke="#00ff00">
+            HTML,
+            Circle::tag()->stroke('#00ff00')->render(),
+            "Failed asserting that element renders correctly with 'stroke' attribute.",
+        );
+    }
+
+    public function testRenderWithStrokeDashArray(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <circle stroke-dasharray="5,5">
+            HTML,
+            Circle::tag()->strokeDashArray('5,5')->render(),
+            "Failed asserting that element renders correctly with 'stroke-dasharray' attribute.",
+        );
+    }
+
+    public function testRenderWithStrokeLineCap(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <circle stroke-linecap="round">
+            HTML,
+            Circle::tag()->strokeLineCap('round')->render(),
+            "Failed asserting that element renders correctly with 'stroke-linecap' attribute.",
+        );
+    }
+
+    public function testRenderWithStrokeLineCapUsingEnum(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <circle stroke-linecap="square">
+            HTML,
+            Circle::tag()->strokeLineCap(StrokeLineCap::SQUARE)->render(),
+            "Failed asserting that element renders correctly with 'stroke-linecap' attribute.",
+        );
+    }
+
+    public function testRenderWithStrokeLineJoin(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <circle stroke-linejoin="round">
+            HTML,
+            Circle::tag()->strokeLineJoin('round')->render(),
+            "Failed asserting that element renders correctly with 'stroke-linejoin' attribute.",
+        );
+    }
+
+    public function testRenderWithStrokeLineJoinUsingEnum(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <circle stroke-linejoin="round">
+            HTML,
+            Circle::tag()->strokeLineJoin(StrokeLineJoin::ROUND)->render(),
+            "Failed asserting that element renders correctly with 'stroke-linejoin' attribute.",
+        );
+    }
+
+    public function testRenderWithStrokeMiterlimit(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <circle stroke-miterlimit="10">
+            HTML,
+            Circle::tag()->strokeMiterlimit(10)->render(),
+            "Failed asserting that element renders correctly with 'stroke-miterlimit' attribute.",
+        );
+    }
+
+    public function testRenderWithStrokeOpacity(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <circle stroke-opacity="0.8">
+            HTML,
+            Circle::tag()->strokeOpacity(0.8)->render(),
+            "Failed asserting that element renders correctly with 'stroke-opacity' attribute.",
+        );
+    }
+
+    public function testRenderWithStrokeWidth(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <circle stroke-width="2">
+            HTML,
+            Circle::tag()->strokeWidth('2')->render(),
+            "Failed asserting that element renders correctly with 'stroke-width' attribute.",
+        );
+    }
+
+    public function testRenderWithStyle(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <circle style='test-value'>
+            HTML,
+            Circle::tag()->style('test-value')->render(),
+            "Failed asserting that element renders correctly with 'style' attribute.",
+        );
+    }
+
+    public function testRenderWithToString(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <circle>
+            HTML,
+            (string) Circle::tag(),
+            "Failed asserting that '__toString()' method renders correctly.",
+        );
+    }
+
+    public function testRenderWithTransform(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <circle transform="rotate(45)">
+            HTML,
+            Circle::tag()->transform('rotate(45)')->render(),
+            "Failed asserting that element renders correctly with 'transform' attribute.",
+        );
+    }
+
+    public function testRenderWithUserOverridesGlobalDefaults(): void
+    {
+        SimpleFactory::setDefaults(Circle::class, ['class' => 'from-global', 'id' => 'id-global']);
+
+        self::equalsWithoutLE(
+            <<<HTML
+            <circle class="from-global" id="id-user">
+            HTML,
+            Circle::tag(['id' => 'id-user'])->render(),
+            'Failed asserting that user-defined attributes override global defaults correctly.',
+        );
+
+        SimpleFactory::setDefaults(Circle::class, []);
+    }
+
+    public function testReturnNewInstanceWhenSettingAttribute(): void
+    {
+        $circle = Circle::tag();
+
+        self::assertNotSame(
+            $circle,
+            $circle->fill(''),
+            'Should return a new instance when setting the attribute, ensuring immutability.',
+        );
+        self::assertNotSame(
+            $circle,
+            $circle->fillOpacity('0'),
+            'Should return a new instance when setting the attribute, ensuring immutability.',
+        );
+        self::assertNotSame(
+            $circle,
+            $circle->fillRule(''),
+            'Should return a new instance when setting the attribute, ensuring immutability.',
+        );
+        self::assertNotSame(
+            $circle,
+            $circle->opacity('0'),
+            'Should return a new instance when setting the attribute, ensuring immutability.',
+        );
+        self::assertNotSame(
+            $circle,
+            $circle->stroke(''),
+            'Should return a new instance when setting the attribute, ensuring immutability.',
+        );
+        self::assertNotSame(
+            $circle,
+            $circle->strokeDashArray(''),
+            'Should return a new instance when setting the attribute, ensuring immutability.',
+        );
+        self::assertNotSame(
+            $circle,
+            $circle->strokeLineCap(''),
+            'Should return a new instance when setting the attribute, ensuring immutability.',
+        );
+        self::assertNotSame(
+            $circle,
+            $circle->strokeLineJoin(''),
+            'Should return a new instance when setting the attribute, ensuring immutability.',
+        );
+        self::assertNotSame(
+            $circle,
+            $circle->strokeMiterlimit('1'),
+            'Should return a new instance when setting the attribute, ensuring immutability.',
+        );
+        self::assertNotSame(
+            $circle,
+            $circle->strokeOpacity('0'),
+            'Should return a new instance when setting the attribute, ensuring immutability.',
+        );
+        self::assertNotSame(
+            $circle,
+            $circle->strokeWidth(''),
+            'Should return a new instance when setting the attribute, ensuring immutability.',
+        );
+        self::assertNotSame(
+            $circle,
+            $circle->transform(''),
+            'Should return a new instance when setting the attribute, ensuring immutability.',
+        );
+    }
+}